Make the wasm feature support the usual APIs
With the fs option, from_path can now be meaningful on a target like
wasm which has no native file system, so reinstate it.

Then for from_string: having it return a different type with the wasm
feature is a bit unfriendly, and quite unnecessary. We can still expose
it in this JS-friendly way while supporting the full error types if
someone is using grass from the Rust side.